fixed bug of latest date algorithm

+++ forestmask_edgedensity.R
@@ -17,23 +17,25 @@ for(file in files){
     if(file.exists(elefile)){
       loss=rast(file)
       masknames=maskboundaries[ext(loss)]$names
-      if(length(masknames)>1){
-        mask=crop(merge(sprc(masknames)),ext(loss))
-      }else{mask=rast(masknames,win=ext(loss))}
-      if(ext(mask)!=ext(loss)){mask=extend(mask,loss,fill=0)}
-      mask=mask-classify(loss,rclmat,others=0)
-      template=rast(paste0("C:/users/jonas/downloads/elevations/",loc,"_elevation.tif"))
-      forestmask=paste0("C:/users/jonas/downloads/elevations/",loc,"_forestmask2019.tif")
-      if(!file.exists(forestmask)){project(mask,template,method="sum",filename=forestmask)}
-      selpols=as.polygons(rast(ext(mask),ncol=2,nrow=2))
-      for(i in 1:length(selpols)){
-        buffsize=res(mask)[1]*110000
-        crop(mask,buffer(selpols[i],buffsize),filename="C:/data/tempras.tif",overwrite=T)
-        system(paste0("python C:/data/git/ForestForesight/sobelfilter.py C:/data/tempras.tif C:/data/edges",i,".tif"),intern=T)
+      if(length(masknames)>0){
+        if(length(masknames)>1){
+          mask=crop(merge(sprc(masknames)),ext(loss))
+        }else{mask=rast(masknames,win=ext(loss))}
+        if(ext(mask)!=ext(loss)){mask=extend(mask,loss,fill=0)}
+        mask=mask-classify(loss,rclmat,others=0)
+        template=rast(paste0("C:/users/jonas/downloads/elevations/",loc,"_elevation.tif"))
+        forestmask=paste0("C:/users/jonas/downloads/elevations/",loc,"_forestmask2019.tif")
+        if(!file.exists(forestmask)){project(mask,template,method="sum",filename=forestmask)}
+        selpols=as.polygons(rast(ext(mask),ncol=2,nrow=2))
+        for(i in 1:length(selpols)){
+          buffsize=res(mask)[1]*110000
+          crop(mask,buffer(selpols[i],buffsize),filename="C:/data/tempras.tif",overwrite=T)
+          system(paste0("python C:/data/git/ForestForesight/sobelfilter.py C:/data/tempras.tif C:/data/edges",i,".tif"),intern=T)
+        }
+        rasts=list.files("C:/data/",full.names = T,pattern="edges")
+        newrast=project(merge(sprc(rasts)),template,method="sum",filename=edgesfile)
+        file.remove(rasts)
       }
-      rasts=list.files("C:/data/",full.names = T,pattern="edges")
-      newrast=project(merge(sprc(rasts)),template,method="sum",filename=edgesfile)
-      file.remove(rasts)
     }
   }
 }
